+++ orca-whirlpool/src/modules/mod.rs
@@ -1,0 +1,5 @@
+#[path = "1_map_block.rs"]
+pub mod map_block;
+
+#[path = "20_graph_out.rs"]
+pub mod graph_out;
